bug 60153: patch from PJ Fanning to demonstrate that SXSSFWorkbook SheetDataWriter can write encrypted temporary files to disk and the workbook can be AES encrypted when written to disk for a fully secure disk environment
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1763969 13f79535-47bb-0310-9956-ffa450edef68

+++ src/ooxml/testcases/org/apache/poi/poifs/crypt/AesZipFileZipEntrySource.java
@@ -69,8 +69,10 @@ public class AesZipFileZipEntrySource implements ZipEntrySource {
 
     @Override
     public void close() throws IOException {
-        zipFile.close();
-        tmpFile.delete();
+        if(!closed) {
+            zipFile.close();
+            tmpFile.delete();
+        }
         closed = true;
     }
